Split common code into separate shared module
