+++ editorconfig.plugin
@@ -1,6 +1,6 @@
 [Plugin]
 Loader=python
-Module=editorconfig
+Module=editorconfig_gedit3
 IAge=3
 Name=EditorConfig
 Description=EditorConfig Gedit plugin
